Removed many unused oauth2 models/classes, and cleaned up user login.
Moved Google Login to UserController.

TODO:NN check the relevant tags and update docs/login functionality
