+++ server/app/controllers/UserController.scala
@@ -1,54 +1,28 @@
-/*
- * Anyplace: A free and open Indoor Navigation Service with superb accuracy!
- *
- * Anyplace is a first-of-a-kind indoor information service offering GPS-less
- * localization, navigation and search inside buildings using ordinary smartphones.
- *
- * Author(s): Nikolas Neofytou, Constantinos Costa, Kyriakos Georgiou, Lambros Petrou
- *
- * Co-Supervisor: Paschalis Mpeis
- * Supervisor: Demetrios Zeinalipour-Yazti
- *
- * URL: https://anyplace.cs.ucy.ac.cy
- * Contact: anyplace@cs.ucy.ac.cy
- *
- * Copyright (c) 2016, Data Management Systems Lab (DMSL), University of Cyprus.
- * All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the “Software”), to deal in the
- * Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
- * and to permit persons to whom the Software is furnished to do so, subject to the
- * following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- *
- */
 package controllers
 
-import datasources.{ProxyDataSource, SCHEMA}
+import datasources.{MongodbDatasource, ProxyDataSource, SCHEMA}
+
+import scala.concurrent.duration.Duration
 import javax.inject.{Inject, Singleton}
 import json.VALIDATE
-import oauth.provider.v2.models.OAuth2Request
+import models.{Account, ExternalType}
+import models.ExternalType.ExternalType
+import models.oauth.OAuth2Request
+import org.mongodb.scala.MongoDatabase
+import org.mongodb.scala.model.Filters.equal
 import play.api.Configuration
 import play.api.libs.json._
 import play.api.mvc._
-import utils.{LOG, RESPONSE}
+import utils.JsonUtils.isNullOrEmpty
+import utils.{JsonUtils, LOG, Network, RESPONSE, Utils}
+
+import scala.concurrent.Await
 
 @Singleton
 class UserController @Inject()(cc: ControllerComponents,
                                 pds: ProxyDataSource,
-                                conf: Configuration)
+                               mongoDB: MongodbDatasource,
+                               conf: Configuration)
   extends AbstractController(cc) {
 
   def login() = Action {
@@ -92,7 +66,9 @@ class UserController @Inject()(cc: ControllerComponents,
         val password = (json \ SCHEMA.fPassword).as[String]
         val external = "anyplace"
         var accType = "user"
-        if (pds.getIDatasource.isAdmin(SCHEMA.cUsers)) // if first user then assign as admin
+
+        // if first user then assign as admin
+        if (pds.getIDatasource.isAdmin(SCHEMA.cUsers))
           accType = "admin"
         // Check if the email is unique
         val storedEmail = pds.getIDatasource.getFromKeyAsJson(SCHEMA.cUsers, SCHEMA.fEmail, email)
@@ -100,15 +76,176 @@ class UserController @Inject()(cc: ControllerComponents,
         // Check if the username is unique
         val storedUsername = pds.getIDatasource.getFromKeyAsJson(SCHEMA.cUsers, SCHEMA.fUsername, username)
         if (storedUsername != null) return RESPONSE.BAD("Username is already taken.")
+
         val newUser = pds.getIDatasource.register(SCHEMA.cUsers, name, email, username, password, external, accType)
         if (newUser == null) return RESPONSE.BAD("Please try again.")
-        // TODO:NN change return to "succesffully ... pls login.."
         val res: JsValue = Json.obj("newUser" -> newUser)
-        return RESPONSE.OK(res,"Succefully registered.")
+        return RESPONSE.OK(res,"Successfully registered.")
       }
       inner(request)
   }
+
+  /**
+   *
+   * @return type(admin, user, .. etc) + message
+   */
+  def loginGoogle() = Action {
+    implicit request =>
+      def inner(request: Request[AnyContent]): Result = {
+        LOG.D1("loginGoogle")
+        val auth = new OAuth2Request(request)
+        if (!auth.assertJsonBody()) return RESPONSE.BAD(RESPONSE.ERROR_JSON_PARSE)
+        var json = auth.getJsonBody()
+        if (isNullOrEmpty(json)) return RESPONSE.BAD(RESPONSE.ERROR_API_USAGE)
+        LOG.D2("JSON BODY: ")
+        LOG.D2(json.toString())
+        json = appendUserType(json) //# TODO auth.appendUserType() // update json directly.. inside auth object..
+        LOG.D2(json.toString())
+        // auth.isGoogleUser() // and hide the below functionality....
+        val external = json \ SCHEMA.fExternal
+        if (external.toOption.isDefined && external.as[String] == "google") {
+          return authorizeGoogleAccount(auth) // auth.addGoogleAccount()
+        } else {
+          // CHECK:NN why will it ever go here?
+          return RESPONSE.BAD(RESPONSE.ERROR_API_USAGE)
+        }
+        //val user: JsValue = Json.obj("user" -> result)
+        //return AnyResponseHelper.ok(user,"ok")
+      }
+
+      inner(request)
+  }
+
+  // TODO if json has not type add type = user
+  def appendUserType(json: JsValue): JsValue = {
+    if ((json \ SCHEMA.fType).toOption.isDefined) {
+      LOG.I("user type exists: " + (json \ SCHEMA.fType).as[String]) // Might crash
+      return json
+    } else {
+      var userType: String = ""
+      if (isFirstUser()) {
+        userType = "admin"
+        LOG.I("Initializing admin user.")
+      } else {
+        LOG.D4("AppendUserType: user")
+        userType = "user"
+      }
+      return json.as[JsObject] + (SCHEMA.fType -> JsString(userType))
+    }
+  }
+
+  def getAccountType(json: JsValue): ExternalType = {
+    val external = json \ SCHEMA.fExternal
+    if (external.toOption.isDefined) {
+      val exts = external.as[String]
+      if (exts == "google") return ExternalType.GOOGLE
+    }
+    ExternalType.LOCAL
+  }
+
+  def isFirstUser(): Boolean = {
+    val mdb: MongoDatabase = mongoDB.getMDB
+    val collection = mdb.getCollection(SCHEMA.cUsers)
+    val users = collection.find()
+    val awaited = Await.result(users.toFuture(), Duration.Inf)
+    val res = awaited.toList
+
+    res.isEmpty
+  }
+
+  def getUser(json: JsValue): JsValue = {
+    // val mdb: MongoDatabase = mongoDB.getMDB  CHECK:NN
+    // val collection = mdb.getCollection(SCHEMA.cUsers)
+    var user: JsValue = null
+    getAccountType(json) match {
+      case ExternalType.GOOGLE =>
+        val mdb: MongoDatabase = mongoDB.getMDB
+        val collection = mdb.getCollection(SCHEMA.cUsers)
+        val ownerId = (json \ SCHEMA.fOwnerId).as[String]
+        val userLookUp = collection.find(equal(SCHEMA.fOwnerId, ownerId))
+        val awaited = Await.result(userLookUp.toFuture(), Duration.Inf)
+        val res = awaited.toList
+        if (res.size == 1) {
+          user = mongoDB.convertJson(res.head) // CHECK:NN was (0)
+        } else if (res.size > 1) {
+          LOG.E("User exists. More than one user with id: " + ownerId)
+        }
+      case ExternalType.LOCAL => LOG.D("TODO: query unique email") // CHECK:NN ??
+    }
+
+    user
+  }
+
+  // TODO:NN explain this.... REVIEW with PM
+  def authorizeGoogleAccount(auth: OAuth2Request): Result = {
+    LOG.I("addGoogleAccount")
+    var json = auth.getJsonBody()
+    LOG.I(json.toString())
+    val hasExternal = JsonUtils.hasProperties(json, SCHEMA.fExternal) // TODO
+    if (!hasExternal.isEmpty) return RESPONSE.MISSING_FIELDS(hasExternal)
+
+    var id = verifyGoogleAuthentication((json \ SCHEMA.fAccessToken).as[String])
+    if (id == null) return RESPONSE.UNAUTHORIZED_USER()
+    id = Utils.appendGoogleIdIfNeeded(id)
+    json = json.as[JsObject] + (SCHEMA.fOwnerId -> Json.toJson(id))
+
+    // if user exists but has no anyplace access_token: create one and put on db
+    var user = getUser(json)
+    val hasAccessToken = !(user \ SCHEMA.fAccessToken).toOption.isEmpty
+    if (hasAccessToken && user != null) { // add access_token to db if !exists
+      val newAccessToken = MongodbDatasource.generateAccessToken(false)
+      user = user.as[JsObject] + (SCHEMA.fAccessToken -> JsString(newAccessToken)) +
+        (SCHEMA.fSchema -> JsNumber(MongodbDatasource.__SCHEMA))
+      pds.getIDatasource.replaceJsonDocument(SCHEMA.cUsers, SCHEMA.fOwnerId,
+        (json \ SCHEMA.fOwnerId).as[String], user.toString())
+    }
+    json = json.as[JsObject] + (SCHEMA.fOwnerId -> Json.toJson(id))
+    if (user != null) { // user and access_token exists
+      user = user.as[JsObject] + (SCHEMA.fType -> JsString((user \ SCHEMA.fType).as[String]))
+      return RESPONSE.OK(user, "User Exists.") // its not AnyResponseHelperok
+    } else {  // new user created CHECK:NN .. how token is on local account?
+      val user = new Account(json)
+      pds.getIDatasource.addJsonDocument(SCHEMA.cUsers, user.toString())
+      return RESPONSE.OK(user.toJson(), "Added new google user.")
+    }
+  }
+
+
+  /**
+   * TODO:NN move from here. Maybe to UserController? (or some Util object?). Does it work with local accounts?
+   * @param authToken
+   * @return
+   */
+  def verifyGoogleAuthentication(authToken: String): String = {
+    // remove the double string quotes due to json processing
+    val gURL = "https://www.googleapis.com/oauth2/v3/tokeninfo?id_token=" + authToken
+    var res = ""
+    try {
+      res = Network.GET(gURL)
+    } catch {
+      case e: Exception => { LOG.E("verifyId", e)}
+    }
+    if (res != null) {
+      try {
+        // CHECK:PM CHECK:NN bug on main branch (JsonObject.fromJson())
+        val json = Json.parse(res)
+        val uid = (json \ "user_id")
+        val sub = (json \ "sub")
+        if (uid.toOption.isDefined)
+          return uid.as[String]
+        if (sub.toOption.isDefined)
+          return sub.as[String]
+      } catch {
+        case iae: IllegalArgumentException => LOG.E("verifyId: " + iae.getMessage + "String: '" + res + "'");
+        case e: Exception => LOG.E("verifyId", e)
+      }
+    }
+    null
+  }
+
+
 }
+
 
 // CHECK:NN
 /**
