--- server/app/oauth/provider/v2/granttype/IGrantHandler.java
@@ -1,59 +1,0 @@
-/*
- * AnyPlace: A free and open Indoor Navigation Service with superb accuracy!
- *
- * Anyplace is a first-of-a-kind indoor information service offering GPS-less
- * localization, navigation and search inside buildings using ordinary smartphones.
- *
- * Author(s): Constantinos Costa, Kyriakos Georgiou, Lambros Petrou
- *
- * Supervisor: Demetrios Zeinalipour-Yazti
- *
- * URL: https://anyplace.cs.ucy.ac.cy
- * Contact: anyplace@cs.ucy.ac.cy
- *
- * Copyright (c) 2016, Data Management Systems Lab (DMSL), University of Cyprus.
- * All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the “Software”), to deal in the
- * Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
- * and to permit persons to whom the Software is furnished to do so, subject to the
- * following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- *
- */
-package oauth.provider.v2.granttype;
-
-import oauth.provider.v2.models.AccountModel;
-import accounts.IAccountService;
-import oauth.provider.v2.models.OAuth2Request;
-import play.mvc.Result;
-
-public interface IGrantHandler {
-
-    /**
-     * Handle a request to issue a token and issue it.
-     * This method should be implemented for each grant type of OAuth2
-     * specification. For instance, the procedure uses an AccountService
-     * instance to access the accounts database. Each grant type has
-     * some validation rule and according to the request satisfying those rules
-     * either a valid token will be returned or an error.
-     *
-     * @param request The current request being processed
-     * @param accountService The instance of the Accounts Service database
-     * @return a Result containing the token or the error
-     */
-    public Result handleRequest(OAuth2Request request, IAccountService accountService, AccountModel account);
-
-}
